vm_lb/requirements.txt
numpy>=1.24
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
